Add code to check keyboardFrame
